Edit src structure and separate Button's interfaces
